requirements.txt
psycopg2-binary>=2.9
dbt-postgres>=1.7,<2.0
